Cut out definition of  "Settings", from GameConfig.swift  to its own file
