+++ Shuffle100/models/GameConfig.swift
@@ -27,32 +27,3 @@ struct GameConfig {
         self.fakeMode = fakeMode
     }
 }
-
-class Settings {
-    var mode: GameConfig
-    var selectedStatus100: SelectedState100
-    
-    init(mode: GameConfig = GameConfig(), bool100: Bool100 = Bool100()){
-        self.mode = mode
-        self.selectedStatus100 = SelectedState100(bool100: bool100)
-    }
-    
-    var reciteMode: ReciteMode {
-        get {
-            return mode.reciteMode
-        }
-        set(m) {
-            mode.reciteMode = m
-        }
-    }
-    
-    var fakeMode: Bool {
-        get {
-            return mode.fakeMode
-        }
-        set(m) {
-            mode.fakeMode = m
-        }
-    }
-}
-
